+++ src/Icon/ArrowUp.js
@@ -1,0 +1,6 @@
+import React from 'react'
+
+const ArrowUp = () => <path
+  d="M14 2l10.706 9.292a1 1 0 01-1.318 1.5L15 5.5V25a1 1 0 01-1.993.117L13 5.5l-8.292 7.206c-.36.361-1.037.406-1.408.02s-.396-.945-.091-1.338L14 2z" />
+
+export default ArrowUp